Switch from Interpolations to DataInterpolations
Primarily because DataInterpolations is already set up to work with MTK, with derivatives etc. etc. already hooked up.  The tests fail because the coefficients are not the same, but from what I can tell the new derivatives are actually correct, so would be good to get some context on what these are actually testing.

My usecase here is to hook up SatelliteToolboxTransformations to some MTK/Dyad blocks so that you can graphically describe your satellite system, and go between different coordinate systems easily.